refactor: eventModal 재사용 가능하도록 개선
