+++ frontend/src/components/specific/modal/eventModal/EventModal.styles.ts
